#111 Made the implementation of the Pathlike generic
