+++ test/unit/conftest.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 from typing import Iterable, ByteString, BinaryIO
-from glob import glob
 import os
 from io import IOBase
 import shutil
 import errno
 from pathlib import Path
+import pytest
 
 
 class BucketFake:
@@ -23,7 +23,10 @@ class BucketFake:
 
     @property
     def files(self) -> list[str]:
-        return glob('**/*.*', root_dir=self.root, recursive=True)
+        root_length = len(str(self.root))
+        if self.root != self.root.root:
+            root_length += 1
+        return [str(pth)[root_length:] for pth in self.root.rglob('*.*')]
 
     def delete(self, path: str) -> None:
         try:
@@ -46,13 +49,25 @@ class BucketFake:
 
     def download(self, path: str, chunk_size: int) -> Iterable[ByteString]:
         full_path = self._get_full_path(path)
-        if full_path.exists():
-            if full_path.is_file():
-                with full_path.open('rb') as f:
-                    while True:
-                        data = f.read(chunk_size)
-                        if data:
-                            yield data
-                        else:
-                            return
-        raise FileNotFoundError(errno.ENOENT, os.strerror(errno.ENOENT), str(path))
+        if (not full_path.exists()) or (not full_path.is_file()):
+            raise FileNotFoundError(errno.ENOENT, os.strerror(errno.ENOENT), str(path))
+        with full_path.open('rb') as f:
+            while True:
+                data = f.read(chunk_size)
+                if not data:
+                    break
+                yield data
+
+
+@pytest.fixture
+def bucket_fake(tmpdir) -> BucketFake:
+    dir1 = tmpdir.mkdir('dir1')
+    dir2 = tmpdir.mkdir('dir2')
+    dir11 = dir1.mkdir('dir11')
+    dir12 = dir1.mkdir('dir12')
+    for d, d_id in zip([tmpdir, dir1, dir2, dir11, dir12], [0, 1, 2, 11, 12]):
+        for i in range(2):
+            file_xx = d / f'file{d_id}{i}.dat'
+            dat = bytes([d_id * i] * 24)
+            file_xx.write_binary(dat)
+    return BucketFake(tmpdir)
