+++ test/unit/test_bucket_path.py
@@ -1,23 +1,7 @@
 from pathlib import Path
 from itertools import chain
 import pytest
-from exasol.bucketfs.bucket_path import BucketPath
-from unit.bucket_fake import BucketFake
-
-
-@pytest.fixture
-def api_mock(tmpdir) -> BucketFake:
-    dir1 = tmpdir.mkdir('dir1')
-    dir2 = tmpdir.mkdir('dir2')
-    dir11 = dir1.mkdir('dir11')
-    dir12 = dir1.mkdir('dir12')
-    for d, d_id in zip([tmpdir, dir1, dir2, dir11, dir12], [0, 1, 2, 11, 12]):
-        for i in range(2):
-            file_name = f'file{d_id}{i}.dat'
-            dat = bytes([d_id * i] * 24)
-            with open(str(d / file_name), 'wb') as f:
-                f.write(dat)
-    return BucketFake(tmpdir)
+from exasol.bucketfs._path import BucketPath
 
 
 @pytest.mark.parametrize("test_path, should_exist", [
@@ -26,8 +10,8 @@ def api_mock(tmpdir) -> BucketFake:
     ('dir1/file19.dat', False),
     ('dir1/dir3', False)
 ])
-def test_file_exists(api_mock, test_path, should_exist):
-    path = BucketPath(test_path, bucket_api=api_mock)
+def test_file_exists(bucket_fake, test_path, should_exist):
+    path = BucketPath(test_path, bucket_api=bucket_fake)
     assert path.exists() == should_exist
 
 
@@ -37,8 +21,8 @@ def test_file_exists(api_mock, test_path, should_exist):
     ('dir1/file19.dat', False),
     ('dir1/dir3', False)
 ])
-def test_is_dir(api_mock, test_path, is_dir):
-    path = BucketPath(test_path, bucket_api=api_mock)
+def test_is_dir(bucket_fake, test_path, is_dir):
+    path = BucketPath(test_path, bucket_api=bucket_fake)
     assert path.is_dir() == is_dir
 
 
@@ -48,82 +32,82 @@ def test_is_dir(api_mock, test_path, is_dir):
     ('dir1/file19.dat', False),
     ('dir1/dir3', False)
 ])
-def test_is_file(api_mock, test_path, is_file):
-    path = BucketPath(test_path, bucket_api=api_mock)
+def test_is_file(bucket_fake, test_path, is_file):
+    path = BucketPath(test_path, bucket_api=bucket_fake)
     assert path.is_file() == is_file
 
 
-def test_rm(api_mock):
-    path = BucketPath('dir1/dir12/file120.dat', bucket_api=api_mock)
+def test_rm(bucket_fake):
+    path = BucketPath('dir1/dir12/file120.dat', bucket_api=bucket_fake)
     path.rm()
     assert not path.exists()
 
 
-def test_rm_not_exist(api_mock):
-    path = BucketPath('dir1/dir12/file125.dat', bucket_api=api_mock)
+def test_rm_not_exist(bucket_fake):
+    path = BucketPath('dir1/dir12/file125.dat', bucket_api=bucket_fake)
     with pytest.raises(FileNotFoundError):
         path.rm()
 
 
-def test_rm_directory(api_mock):
-    path = BucketPath('dir1/dir12', bucket_api=api_mock)
+def test_rm_directory(bucket_fake):
+    path = BucketPath('dir1/dir12', bucket_api=bucket_fake)
     with pytest.raises(IsADirectoryError):
         path.rm()
 
 
-def test_rmdir(api_mock):
+def test_rmdir(bucket_fake):
     for i in range(2):
-        BucketPath(f'dir1/dir12/file12{i}.dat', bucket_api=api_mock).rm()
-    path = BucketPath('dir1/dir12', bucket_api=api_mock)
+        BucketPath(f'dir1/dir12/file12{i}.dat', bucket_api=bucket_fake).rm()
+    path = BucketPath('dir1/dir12', bucket_api=bucket_fake)
     path.rmdir(recursive=False)
     assert not path.exists()
 
 
-def test_rmdir_recursive(api_mock):
-    path = BucketPath('dir1', bucket_api=api_mock)
+def test_rmdir_recursive(bucket_fake):
+    path = BucketPath('dir1', bucket_api=bucket_fake)
     path.rmdir(recursive=True)
     assert not path.exists()
 
 
-def test_rmdir_not_empty(api_mock):
-    path = BucketPath('dir1', bucket_api=api_mock)
+def test_rmdir_not_empty(bucket_fake):
+    path = BucketPath('dir1', bucket_api=bucket_fake)
     with pytest.raises(OSError):
         path.rmdir(recursive=False)
 
 
-def test_rmdir_not_exist(api_mock):
-    path = BucketPath('dir1/dir5', bucket_api=api_mock)
+def test_rmdir_not_exist(bucket_fake):
+    path = BucketPath('dir1/dir5', bucket_api=bucket_fake)
     path.rmdir()
 
 
-def test_rmdir_file(api_mock):
-    path = BucketPath('dir1/dir12/file120.dat', bucket_api=api_mock)
+def test_rmdir_file(bucket_fake):
+    path = BucketPath('dir1/dir12/file120.dat', bucket_api=bucket_fake)
     with pytest.raises(NotADirectoryError):
         path.rmdir()
 
 
-def test_joinpath(api_mock):
-    path1 = BucketPath('dir1', bucket_api=api_mock)
+def test_joinpath(bucket_fake):
+    path1 = BucketPath('dir1', bucket_api=bucket_fake)
     path2 = 'dir11'
-    path3 = BucketPath('dir111/dir1111', bucket_api=api_mock)
+    path3 = BucketPath('dir111/dir1111', bucket_api=bucket_fake)
     path4 = Path('dir11111/file111110.dat')
     path = path1.joinpath(path2, path3, path4)
     assert isinstance(path, BucketPath)
     assert str(path) == 'dir1/dir11/dir111/dir1111/dir11111/file111110.dat'
 
 
-def test_truediv(api_mock):
-    path1 = BucketPath('dir1', bucket_api=api_mock)
+def test_truediv(bucket_fake):
+    path1 = BucketPath('dir1', bucket_api=bucket_fake)
     path2 = 'dir11'
-    path3 = BucketPath('dir111/dir1111', bucket_api=api_mock)
+    path3 = BucketPath('dir111/dir1111', bucket_api=bucket_fake)
     path4 = Path('dir11111/file111110.dat')
     path = path1 / path2 / path3 / path4
     assert isinstance(path, BucketPath)
     assert str(path) == 'dir1/dir11/dir111/dir1111/dir11111/file111110.dat'
 
 
-def test_walk_top_down(api_mock):
-    path = BucketPath('', bucket_api=api_mock)
+def test_walk_top_down(bucket_fake):
+    path = BucketPath('', bucket_api=bucket_fake)
     content = [','.join(chain([pth.name, '/'], sorted(dirs), sorted(files)))
                for pth, dirs, files in path.walk(top_down=True)]
     expected_content = [
@@ -138,8 +122,8 @@ def test_walk_top_down(api_mock):
     assert idx == sorted(idx)
 
 
-def test_walk_bottom_up(api_mock):
-    path = BucketPath('', bucket_api=api_mock)
+def test_walk_bottom_up(bucket_fake):
+    path = BucketPath('', bucket_api=bucket_fake)
     content = [','.join(chain([pth.name, '/'], sorted(dirs), sorted(files)))
                for pth, dirs, files in path.walk(top_down=False)]
     expected_content = [
@@ -154,8 +138,8 @@ def test_walk_bottom_up(api_mock):
     assert idx == sorted(idx)
 
 
-def test_iterdir(api_mock):
-    path = BucketPath('dir1', bucket_api=api_mock)
+def test_iterdir(bucket_fake):
+    path = BucketPath('dir1', bucket_api=bucket_fake)
     content = set(str(node) for node in path.iterdir())
     expected_content = {
         'dir1/dir11',
@@ -166,49 +150,49 @@ def test_iterdir(api_mock):
     assert content == expected_content
 
 
-def test_read(api_mock):
-    path = BucketPath('dir1/dir12/file121.dat', bucket_api=api_mock)
+def test_read(bucket_fake):
+    path = BucketPath('dir1/dir12/file121.dat', bucket_api=bucket_fake)
     expected_chunk = bytes([12] * 8)
     for chunk in path.read(chunk_size=8):
         assert chunk == expected_chunk
 
 
-def test_read_not_found(api_mock):
-    path = BucketPath('dir1/file12.dat', bucket_api=api_mock)
+def test_read_not_found(bucket_fake):
+    path = BucketPath('dir1/file12.dat', bucket_api=bucket_fake)
     with pytest.raises(FileNotFoundError):
         list(path.read())
 
 
 @pytest.mark.parametrize("file_name", ['file23.dat', 'file20.dat'])
-def test_write_bytes(api_mock, file_name):
+def test_write_bytes(bucket_fake, file_name):
     data = b'abcd'
-    path = BucketPath(f'dir2/{file_name}', bucket_api=api_mock)
+    path = BucketPath(f'dir2/{file_name}', bucket_api=bucket_fake)
     path.write(data)
     data_back = next(iter(path.read(100)))
     assert data_back == data
 
 
-def test_write_chunks(api_mock):
+def test_write_chunks(bucket_fake):
     data_chunks = [b'abc', b'def', b'gh']
-    path = BucketPath('dir2/file23.dat', bucket_api=api_mock)
+    path = BucketPath('dir2/file23.dat', bucket_api=bucket_fake)
     path.write(data_chunks)
     data_back = next(iter(path.read(100)))
     assert data_back == b'abcdefgh'
 
 
-def test_write_file(api_mock):
-    path = BucketPath('dir2/file_copy.dat', bucket_api=api_mock)
-    source_file = api_mock.root / 'dir2/file21.dat'
+def test_write_file(bucket_fake):
+    path = BucketPath('dir2/file_copy.dat', bucket_api=bucket_fake)
+    source_file = bucket_fake.root / 'dir2/file21.dat'
     with open(source_file, 'rb') as f:
         path.write(f)
     with open(source_file, 'rb') as f:
         assert next(iter(path.read(100))) == f.read()
 
 
-def test_write_and_create_parent(api_mock):
-    path = BucketPath('dir2/dir21/file_copy.dat', bucket_api=api_mock)
+def test_write_and_create_parent(bucket_fake):
+    path = BucketPath('dir2/dir21/file_copy.dat', bucket_api=bucket_fake)
     assert not path.exists()
-    source_file = api_mock.root / 'dir2/file21.dat'
+    source_file = bucket_fake.root / 'dir2/file21.dat'
     with open(source_file, 'rb') as f:
         path.write(f)
     assert path.exists()
